rename reducers
